+++ src/components/v2/V2Project/V2ProjectSettings/index.ts
@@ -1,0 +1,1 @@
+export { V2ProjectSettings } from './V2ProjectSettings'
