+++ src/components/v2/V2Project/V2ProjectSettings/pages/V2ProjectDetails.tsx
@@ -1,15 +1,18 @@
-import { Trans } from '@lingui/macro'
 import { useForm } from 'antd/lib/form/Form'
 import ProjectDetailsForm, {
   ProjectDetailsFormFields,
 } from 'components/forms/ProjectDetailsForm'
-import { PROJECT_PAY_CHARACTER_LIMIT } from 'constants/numbers'
 import { ThemeContext } from 'contexts/themeContext'
+
 import { V2ProjectContext } from 'contexts/v2/projectContext'
 import { useEditV2ProjectDetailsTx } from 'hooks/v2/transactor/EditV2ProjectDetailsTx'
 import React, { useCallback, useContext, useEffect, useState } from 'react'
 import { uploadProjectMetadata } from 'utils/ipfs'
 import { revalidateProject } from 'utils/revalidateProject'
+
+import { Trans } from '@lingui/macro'
+
+import { PROJECT_PAY_CHARACTER_LIMIT } from 'constants/numbers'
 
 const V2ProjectDetails = () => {
   const { colors } = useContext(ThemeContext).theme
