+++ src/pages/v2/p/[projectId]/settings/index.page.tsx
@@ -1,5 +1,5 @@
 import { AppWrapper } from 'components/common'
-import V2ProjectSettings from 'components/v2/V2Project/V2ProjectSettings/V2ProjectSettings'
+import { V2ProjectSettings } from 'components/v2/V2Project/V2ProjectSettings'
 import { ProjectMetadataV4 } from 'models/project-metadata'
 import { GetServerSideProps } from 'next'
 import { V2UserProvider } from 'providers/v2/UserProvider'
